Revert "Add error detection flow that validates node module imports progressi…"
This reverts commit b60f008697cf70c2b80aa11f4d554f7d5c035d6b.
