+++ tests/features/throw-error-undeclared-nested-dep.test.tsx
@@ -30,7 +30,5 @@ test("should throw error when nested import uses undeclared dependency", async (
         `,
       },
     }),
-  ).rejects.toThrow(
-    /Node module imported but not in package\.json "adom-library"/,
-  )
+  ).rejects.toThrow(/Package "adom-library" is not declared in package\.json/)
 })
